move search code from the data-access-plugin to search-access
