+++ packages/data-access-plugin/package.json
@@ -38,8 +38,6 @@
         "elasticsearch-store": "^0.6.0",
         "graphql": "^14.2.1",
         "graphql-type-json": "^0.3.0",
-        "lodash.get": "^4.4.2",
-        "lodash.set": "^4.3.2",
         "terafoundation": "^0.9.0",
         "xlucene-evaluator": "^0.7.0"
     },
@@ -48,8 +46,6 @@
         "@types/express": "^4.16.1",
         "@types/got": "^9.4.1",
         "@types/graphql-type-json": "^0.1.3",
-        "@types/lodash.get": "^4.4.6",
-        "@types/lodash.set": "^4.3.6",
         "elasticsearch": "^15.4.1",
         "express": "^4.16.4",
         "got": "^9.6.0",
