+++ packages/utils/package.json
@@ -32,9 +32,13 @@
         "debug": "^4.1.1",
         "is-plain-object": "^2.0.4",
         "kind-of": "^6.0.2",
-        "lodash.clonedeep": "^4.5.0"
+        "lodash.clonedeep": "^4.5.0",
+        "lodash.get": "^4.4.2",
+        "lodash.set": "^4.3.2"
     },
     "devDependencies": {
+        "@types/lodash.get": "^4.4.6",
+        "@types/lodash.set": "^4.3.6",
         "benchmark": "^2.1.4"
     }
 }
